Refactor recursive load_json_pages() calls to be iterative
Python doesn't do tail call optimisation and this becomes an issue when
downloading the datasets themselves.
Many datasets have hundreds of pages worth of data (623 in the case that
caused the out-of-memory error). This means that all the decoded JSON ends
up "on the stack" and things get really slow because there's lots of active
memory that must be moved during each GC.

Eventually the process grows enough that the Python interpreter terminates
the execution with a MemoryError exception.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>