Institute some timeouts for the HTTP Requests
Sometimes the StatsWales2 OData infrastructure hangs after the HTTP
Request is sent but before any Response is received. When this happens
no Response is ever received and the Request must be aborted and
restarted.

Requests seem to hang for a period of time at least once a day in the
morning, possibly around the time when new cubes are being published. It
also happens sporadically at other points throughout the day.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>