Add check_nulls arguments to CHECK_ROM()
check_nulls is a performance hack as the (%s = ? OR %s IS ?) clause is
expensive on large tables because it cannot take advantage of indexes
with compound keys. But in the case of dataset_dimension which has all
the columns as PRIMARY KEYS, we know that there cannot be any NULLs so
can set check_nulls to False and still obtain correct behaviour.

The (%s = ? OR %s IS ?) clause cause loads into other tables to slow
down as the table size grows but they are small enough that it is not a
problem that needs to be addressed at the moment.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>