Teach load_datasets() how to restart loading from a specific place in the list of datasets
This is useful if the spider has to be restarted halfway thru' a load
because otherwise load_datasets() will spool thru' the list from the
beginning, at best reparsing all the cached JSON and at worst actually
trying to load the data and getting constraint violations for already
loaded datasets.

As load_dataset() loads each dataset inside its own transaction, it should
always be safe to start loading an interrupted dataset from the beginning,
however, preserving the spider_uri_cache from an aborted load is left as an
exercise for the operator. Doing it automatically is left to future work.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>