Summarise datasets by number of dimensions per dataset
A histogram that shows how many datasets have each number of dimensions.

The question of the maximum number of dimensions required by a dataset came
up in Standup on Monday (2024/06/17) morning. To provide an answer I wrote
this SQL query so we may as well keep it in case we want to talk about it
again.

It was mentioned that StatsWales 2 does not perform well when datasets have
more than 10 dimensions so there may be datasets that could benefit from
having more dimensions but they do not show up in this query because they
will have been published as separate cubes.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>